+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v2.17.0
   * `Empty` - this specification filters for collections using `is empty` or `is not empty`, depending on the value of the parameter passed in (e.g. ` where customer.orders is empty`).
   * `isTrue`, `isFalse` - these specifications filter with `true`/`false` value of particular field defined under `path` in `@Spec` annotation.
   * `True` - this specification filters using `true` or `false` for a boolean type field, depending on the value of the parameter passed in.
+  * `isMember`, `isNotMember` - filters with particular member in collection defined under `path` in `@Spec` annotation.
 
 v2.16.0
 =======